refactor(firestore): Remove types functions
