+++ firestore/pipeline_filter_condition.go
@@ -303,51 +303,3 @@ func StartsWith(exprOrFieldPath any, prefix any) BooleanExpr {
 func StringContains(exprOrFieldPath any, substring any) BooleanExpr {
 	return &baseBooleanExpr{baseFunction: newBaseFunction("string_contains", []Expr{asFieldExpr(exprOrFieldPath), asStringExpr(substring)})}
 }
-
-// IsNaN creates a boolean expression that checks if a field or expression evaluates to NaN.
-//
-// exprOrFieldPath is the field path string, [FieldPath] or [Expr] that will be evaluated.
-//
-// Example:
-//
-//	// Check if the 'score' field is NaN.
-//	IsNaN("score")
-func IsNaN(exprOrFieldPath any) BooleanExpr {
-	return &baseBooleanExpr{baseFunction: newBaseFunction("is_nan", []Expr{asFieldExpr(exprOrFieldPath)})}
-}
-
-// IsNotNaN creates a boolean expression that checks if a field or expression does not evaluate to NaN.
-//
-// exprOrFieldPath is the field path string, [FieldPath] or [Expr] that will be evaluated.
-//
-// Example:
-//
-//	// Check if the 'score' field is not NaN.
-//	IsNotNaN("score")
-func IsNotNaN(exprOrFieldPath any) BooleanExpr {
-	return &baseBooleanExpr{baseFunction: newBaseFunction("is_not_nan", []Expr{asFieldExpr(exprOrFieldPath)})}
-}
-
-// IsNull creates a boolean expression that checks if a field or expression evaluates to null.
-//
-// exprOrFieldPath is the field path string, [FieldPath] or [Expr] that will be evaluated.
-//
-// Example:
-//
-//	// Check if the 'address' field is null.
-//	IsNull("address")
-func IsNull(exprOrFieldPath any) BooleanExpr {
-	return &baseBooleanExpr{baseFunction: newBaseFunction("is_null", []Expr{asFieldExpr(exprOrFieldPath)})}
-}
-
-// IsNotNull creates a boolean expression that checks if a field or expression does not evaluate to null.
-//
-// exprOrFieldPath is the field path string, [FieldPath] or [Expr] that will be evaluated.
-//
-// Example:
-//
-//	// Check if the 'address' field is not null.
-//	IsNotNull("address")
-func IsNotNull(exprOrFieldPath any) BooleanExpr {
-	return &baseBooleanExpr{baseFunction: newBaseFunction("is_not_null", []Expr{asFieldExpr(exprOrFieldPath)})}
-}
