+++ firestore/pipeline_function.go
@@ -60,14 +60,6 @@ func newBaseFunction(name string, params []Expr) *baseFunction {
 // Add creates an expression that adds two expressions together, returning it as an Expr.
 // - left can be a field path string, [FieldPath] or [Expr].
 // - right can be a numeric constant or a numeric [Expr].
-//
-// Example:
-//
-//	// Add 5 to the value of the 'age' field.
-//	Add("age", 5)
-//
-//	// Add 'height' to 'weight' field.
-//	Add(FieldOf("height"), FieldOf("weight"))
 func Add(left, right any) Expr {
 	return leftRightToBaseFunction("add", left, right)
 }
@@ -75,14 +67,6 @@ func Add(left, right any) Expr {
 // Subtract creates an expression that subtracts the right expression from the left expression, returning it as an Expr.
 // - left can be a field path string, [FieldPath] or [Expr].
 // - right can be a constant or an [Expr].
-//
-// Example:
-//
-//	// Subtract 5 from the value of the 'age' field.
-//	Subtract("age", 5)
-//
-//	// Subtract 'discount' from 'price' field.
-//	Subtract(FieldOf("price"), FieldOf("discount"))
 func Subtract(left, right any) Expr {
 	return leftRightToBaseFunction("subtract", left, right)
 }
@@ -90,14 +74,6 @@ func Subtract(left, right any) Expr {
 // Multiply creates an expression that multiplies the left and right expressions, returning it as an Expr.
 // - left can be a field path string, [FieldPath] or [Expr].
 // - right can be a constant or an [Expr].
-//
-// Example:
-//
-//	// Multiply 5 to the value of the 'age' field.
-//	Multiply("age", 5)
-//
-//	// Multiply 'discount' and 'price' fields.
-//	Multiply(FieldOf("price"), FieldOf("discount"))
 func Multiply(left, right any) Expr {
 	return leftRightToBaseFunction("multiply", left, right)
 }
@@ -105,58 +81,30 @@ func Multiply(left, right any) Expr {
 // Divide creates an expression that divides the left expression by the right expression, returning it as an Expr.
 // - left can be a field path string, [FieldPath] or [Expr].
 // - right can be a constant or an [Expr].
-//
-// Example:
-//
-//	// Divide the value of the 'age' field by 5.
-//	Divide("age", 5)
-//
-//	// Divide 'discount' field by 'price' field.
-//	Divide(FieldOf("price"), FieldOf("discount"))
 func Divide(left, right any) Expr {
 	return leftRightToBaseFunction("divide", left, right)
 }
 
 // Abs creates an expression that is the absolute value of the input field or expression.
 // - numericExprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that returns a number when evaluated.
-//
-// Example:
-//
-//	// Absolute value of the 'age' field.
-//	Abs("age")
 func Abs(numericExprOrFieldPath any) Expr {
 	return newBaseFunction("abs", []Expr{asFieldExpr(numericExprOrFieldPath)})
 }
 
 // Floor creates an expression that is the largest integer that isn't less than the input field or expression.
 // - numericExprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that returns a number when evaluated.
-//
-// Example:
-//
-//	// Floor value of the 'age' field.
-//	Floor("age")
 func Floor(numericExprOrFieldPath any) Expr {
 	return newBaseFunction("floor", []Expr{asFieldExpr(numericExprOrFieldPath)})
 }
 
 // Ceil creates an expression that is the smallest integer that isn't less than the input field or expression.
 // - numericExprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that returns a number when evaluated.
-//
-// Example:
-//
-//	// Ceiling value of the 'age' field.
-//	Ceil("age")
 func Ceil(numericExprOrFieldPath any) Expr {
 	return newBaseFunction("ceil", []Expr{asFieldExpr(numericExprOrFieldPath)})
 }
 
 // Exp creates an expression that is the Euler's number e raised to the power of the input field or expression.
 // - numericExprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that returns a number when evaluated.
-//
-// Example:
-//
-//	// e to the power of the value of the 'age' field.
-//	Exp("age")
 func Exp(numericExprOrFieldPath any) Expr {
 	return newBaseFunction("exp", []Expr{asFieldExpr(numericExprOrFieldPath)})
 }
@@ -164,36 +112,18 @@ func Exp(numericExprOrFieldPath any) Expr {
 // Log creates an expression that is logarithm of the left expression to base as the right expression, returning it as an Expr.
 // - left can be a field path string, [FieldPath] or [Expr].
 // - right can be a constant or an [Expr].
-//
-// Example:
-//
-//	// Logarithm of 'age' field to base 5.
-//	Log("age", 5)
-//
-//	// Log 'height' to base 'weight' field.
-//	Log(FieldOf("height"), FieldOf("weight"))
 func Log(left, right any) Expr {
 	return leftRightToBaseFunction("log", left, right)
 }
 
 // Log10 creates an expression that is the base 10 logarithm of the input field or expression.
 // - numericExprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that returns a number when evaluated.
-//
-// Example:
-//
-//	// Base 10 logarithmic value of the 'age' field.
-//	Log10("age")
 func Log10(numericExprOrFieldPath any) Expr {
 	return newBaseFunction("log10", []Expr{asFieldExpr(numericExprOrFieldPath)})
 }
 
 // Ln creates an expression that is the natural logarithm (base e) of the input field or expression.
 // - numericExprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that returns a number when evaluated.
-//
-// Example:
-//
-//	// Natural logarithmic value of the 'age' field.
-//	Ln("age")
 func Ln(numericExprOrFieldPath any) Expr {
 	return newBaseFunction("ln", []Expr{asFieldExpr(numericExprOrFieldPath)})
 }
@@ -201,14 +131,6 @@ func Ln(numericExprOrFieldPath any) Expr {
 // Mod creates an expression that computes the modulo of the left expression by the right expression, returning it as an Expr.
 // - left can be a field path string, [FieldPath] or [Expr].
 // - right can be a constant or an [Expr].
-//
-// Example:
-//
-//	// Modulo of 'age' field by 5.
-//	Mod("age", 5)
-//
-//	// Modulo of 'price' field by 'discount' field.
-//	Mod(FieldOf("price"), FieldOf("discount"))
 func Mod(left, right any) Expr {
 	return leftRightToBaseFunction("mod", left, right)
 }
@@ -216,36 +138,18 @@ func Mod(left, right any) Expr {
 // Pow creates an expression that computes the left expression raised to the power of the right expression, returning it as an Expr.
 // - left can be a field path string, [FieldPath] or [Expr].
 // - right can be a constant or an [Expr].
-//
-// Example:
-//
-//	// 'age' field raised to the power of 5.
-//	Pow("age", 5)
-//
-//	// 'price' field raised to the power of 'discount' field.
-//	Pow(FieldOf("price"), FieldOf("discount"))
 func Pow(left, right any) Expr {
 	return leftRightToBaseFunction("pow", left, right)
 }
 
 // Round creates an expression that rounds the input field or expression to nearest integer.
 // - numericExprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that returns a number when evaluated.
-//
-// Example:
-//
-//	// Round the value of the 'age' field.
-//	Round("age")
 func Round(numericExprOrFieldPath any) Expr {
 	return newBaseFunction("round", []Expr{asFieldExpr(numericExprOrFieldPath)})
 }
 
 // Sqrt creates an expression that is the square root of the input field or expression.
 // - numericExprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that returns a number when evaluated.
-//
-// Example:
-//
-//	// Square root of the value of the 'age' field.
-//	Sqrt("age")
 func Sqrt(numericExprOrFieldPath any) Expr {
 	return newBaseFunction("sqrt", []Expr{asFieldExpr(numericExprOrFieldPath)})
 }
@@ -254,11 +158,6 @@ func Sqrt(numericExprOrFieldPath any) Expr {
 // - timestamp can be a field path string, [FieldPath] or [Expr].
 // - unit can be a string or an [Expr]. Valid units include "microsecond", "millisecond", "second", "minute", "hour" and "day".
 // - amount can be an int, int32, int64 or [Expr].
-//
-// Example:
-//
-//	// Add 5 hours to the value of the 'last_updated' field.
-//	TimestampAdd("last_updated", "hour", 5)
 func TimestampAdd(timestamp, unit, amount any) Expr {
 	return newBaseFunction("timestamp_add", []Expr{asFieldExpr(timestamp), asStringExpr(unit), asInt64Expr(amount)})
 }
@@ -267,11 +166,6 @@ func TimestampAdd(timestamp, unit, amount any) Expr {
 // - timestamp can be a field path string, [FieldPath] or [Expr].
 // - unit can be a string or an [Expr]. Valid units include "microsecond", "millisecond", "second", "minute", "hour" and "day".
 // - amount can be an int, int32, int64 or [Expr].
-//
-// Example:
-//
-//	// Subtract 10 days from the value of the 'last_updated' field.
-//	TimestampSubtract("last_updated", "day", 10)
 func TimestampSubtract(timestamp, unit, amount any) Expr {
 	return newBaseFunction("timestamp_subtract", []Expr{asFieldExpr(timestamp), asStringExpr(unit), asInt64Expr(amount)})
 }
@@ -322,22 +216,12 @@ func CurrentTimestamp() Expr {
 
 // ArrayLength creates an expression that calculates the length of an array.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to an array.
-//
-// Example:
-//
-//	// Get the length of the 'tags' array field.
-//	ArrayLength("tags")
 func ArrayLength(exprOrFieldPath any) Expr {
 	return newBaseFunction("array_length", []Expr{asFieldExpr(exprOrFieldPath)})
 }
 
 // Array creates an expression that represents a Firestore array.
 // - elements can be any number of values or expressions that will form the elements of the array.
-//
-// Example:
-//
-//	// Create an array of numbers.
-//	Array(1, 2, 3)
 func Array(elements ...any) Expr {
 	return newBaseFunction("array", toExprs(elements))
 }
@@ -353,22 +237,12 @@ func ArrayFromSlice[T any](elements []T) Expr {
 // ArrayGet creates an expression that retrieves an element from an array at a specified index.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to an array.
 // - offset is the 0-based index of the element to retrieve.
-//
-// Example:
-//
-//	// Get the first element of the 'tags' array field.
-//	ArrayGet("tags", 0)
 func ArrayGet(exprOrFieldPath any, offset any) Expr {
 	return newBaseFunction("array_get", []Expr{asFieldExpr(exprOrFieldPath), asInt64Expr(offset)})
 }
 
 // ArrayReverse creates an expression that reverses the order of elements in an array.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to an array.
-//
-// Example:
-//
-//	// Reverse the 'tags' array.
-//	ArrayReverse("tags")
 func ArrayReverse(exprOrFieldPath any) Expr {
 	return newBaseFunction("array_reverse", []Expr{asFieldExpr(exprOrFieldPath)})
 }
@@ -376,44 +250,24 @@ func ArrayReverse(exprOrFieldPath any) Expr {
 // ArrayConcat creates an expression that concatenates multiple arrays into a single array.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to an array.
 // - otherArrays are the other arrays to concatenate.
-//
-// Example:
-//
-//	// Concatenate the 'tags' and 'categories' array fields.
-//	ArrayConcat("tags", FieldOf("categories"))
 func ArrayConcat(exprOrFieldPath any, otherArrays ...any) Expr {
 	return newBaseFunction("array_concat", append([]Expr{asFieldExpr(exprOrFieldPath)}, toExprs(otherArrays)...))
 }
 
 // ArraySum creates an expression that calculates the sum of all elements in a numeric array.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a numeric array.
-//
-// Example:
-//
-//	// Calculate the sum of the 'scores' array.
-//	ArraySum("scores")
 func ArraySum(exprOrFieldPath any) Expr {
 	return newBaseFunction("sum", []Expr{asFieldExpr(exprOrFieldPath)})
 }
 
 // ArrayMaximum creates an expression that finds the maximum element in a numeric array.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a numeric array.
-//
-// Example:
-//
-//	// Find the maximum value in the 'scores' array.
-//	ArrayMaximum("scores")
 func ArrayMaximum(exprOrFieldPath any) Expr {
 	return newBaseFunction("maximum", []Expr{asFieldExpr(exprOrFieldPath)})
 }
 
 // ArrayMinimum creates an expression that finds the minimum element in a numeric array.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a numeric array.
-//
-// Example:
-//
-//	// Find the minimum value in the 'scores' array.
-//	ArrayMinimum("scores")
 func ArrayMinimum(exprOrFieldPath any) Expr {
 	return newBaseFunction("minimum", []Expr{asFieldExpr(exprOrFieldPath)})
 }
@@ -421,22 +275,12 @@ func ArrayMinimum(exprOrFieldPath any) Expr {
 // ByteLength creates an expression that calculates the length of a string represented by a field or [Expr] in UTF-8
 // bytes, or just the length of a Blob.
 //   - exprOrFieldPath can be a field path string, [FieldPath] or [Expr].
-//
-// Example:
-//
-//	// Get the byte length of the 'name' field.
-//	ByteLength("name")
 func ByteLength(exprOrFieldPath any) Expr {
 	return newBaseFunction("byte_length", []Expr{asFieldExpr(exprOrFieldPath)})
 }
 
 // CharLength creates an expression that calculates the character length of a string field or expression in UTF8.
 //   - exprOrFieldPath can be a field path string, [FieldPath] or [Expr].
-//
-// Example:
-//
-//	// Get the character length of the 'name' field.
-//	CharLength("name")
 func CharLength(exprOrFieldPath any) Expr {
 	return newBaseFunction("char_length", []Expr{asFieldExpr(exprOrFieldPath)})
 }
@@ -444,22 +288,12 @@ func CharLength(exprOrFieldPath any) Expr {
 // StringConcat creates an expression that concatenates multiple strings into a single string.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a string.
 // - otherStrings are the other strings to concatenate.
-//
-// Example:
-//
-//	// Concatenate first name and last name.
-//	StringConcat(FieldOf("firstName"), " ", FieldOf("lastName"))
 func StringConcat(exprOrFieldPath any, otherStrings ...any) Expr {
 	return newBaseFunction("string_concat", append([]Expr{asFieldExpr(exprOrFieldPath)}, toExprs(otherStrings)...))
 }
 
 // StringReverse creates an expression that reverses a string.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a string.
-//
-// Example:
-//
-//	// Reverse the 'name' field.
-//	StringReverse("name")
 func StringReverse(exprOrFieldPath any) Expr {
 	return newBaseFunction("string_reverse", []Expr{asFieldExpr(exprOrFieldPath)})
 }
@@ -467,11 +301,6 @@ func StringReverse(exprOrFieldPath any) Expr {
 // Join creates an expression that joins the elements of a string array into a single string.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a string array.
 // - separator is the string to use as a separator between elements.
-//
-// Example:
-//
-//	// Join the 'tags' array with a comma and space.
-//	Join("tags", ", ")
 func Join(exprOrFieldPath any, separator any) Expr {
 	return newBaseFunction("join", []Expr{asFieldExpr(exprOrFieldPath), asStringExpr(separator)})
 }
@@ -480,44 +309,24 @@ func Join(exprOrFieldPath any, separator any) Expr {
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a string.
 // - index is the starting index of the substring.
 // - offset is the length of the substring.
-//
-// Example:
-//
-//	// Get the first 5 characters of the 'description' field.
-//	Substring("description", 0, 5)
 func Substring(exprOrFieldPath any, index any, offset any) Expr {
 	return newBaseFunction("substring", []Expr{asFieldExpr(exprOrFieldPath), asInt64Expr(index), asInt64Expr(offset)})
 }
 
 // ToLower creates an expression that converts a string to lowercase.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a string.
-//
-// Example:
-//
-//	// Convert the 'username' to lowercase.
-//	ToLower("username")
 func ToLower(exprOrFieldPath any) Expr {
 	return newBaseFunction("to_lower", []Expr{asFieldExpr(exprOrFieldPath)})
 }
 
 // ToUpper creates an expression that converts a string to uppercase.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a string.
-//
-// Example:
-//
-//	// Convert the 'product_code' to uppercase.
-//	ToUpper("product_code")
 func ToUpper(exprOrFieldPath any) Expr {
 	return newBaseFunction("to_upper", []Expr{asFieldExpr(exprOrFieldPath)})
 }
 
 // Trim creates an expression that removes leading and trailing whitespace from a string.
 // - exprOrFieldPath can be a field path string, [FieldPath] or an [Expr] that evaluates to a string.
-//
-// Example:
-//
-//	// Trim the 'email' field.
-//	Trim("email")
 func Trim(exprOrFieldPath any) Expr {
 	return newBaseFunction("trim", []Expr{asFieldExpr(exprOrFieldPath)})
 }
@@ -525,11 +334,6 @@ func Trim(exprOrFieldPath any) Expr {
 // CosineDistance creates an expression that calculates the cosine distance between two vectors.
 //   - vector1 can be a field path string, [FieldPath] or [Expr].
 //   - vector2 can be [Vector32], [Vector64], []float32, []float64 or [Expr].
-//
-// Example:
-//
-//	// Calculate the cosine distance between two vector fields.
-//	CosineDistance("vector_field_1", FieldOf("vector_field_2"))
 func CosineDistance(vector1 any, vector2 any) Expr {
 	return newBaseFunction("cosine_distance", []Expr{asFieldExpr(vector1), asVectorExpr(vector2)})
 }
@@ -537,11 +341,6 @@ func CosineDistance(vector1 any, vector2 any) Expr {
 // DotProduct creates an expression that calculates the dot product of two vectors.
 //   - vector1 can be a field path string, [FieldPath] or [Expr].
 //   - vector2 can be [Vector32], [Vector64], []float32, []float64 or [Expr].
-//
-// Example:
-//
-//	// Calculate the dot product of two vector fields.
-//	DotProduct("vector_field_1", FieldOf("vector_field_2"))
 func DotProduct(vector1 any, vector2 any) Expr {
 	return newBaseFunction("dot_product", []Expr{asFieldExpr(vector1), asVectorExpr(vector2)})
 }
@@ -549,22 +348,12 @@ func DotProduct(vector1 any, vector2 any) Expr {
 // EuclideanDistance creates an expression that calculates the euclidean distance between two vectors.
 //   - vector1 can be a field path string, [FieldPath] or [Expr].
 //   - vector2 can be [Vector32], [Vector64], []float32, []float64 or [Expr].
-//
-// Example:
-//
-//	// Calculate the euclidean distance between two vector fields.
-//	EuclideanDistance("vector_field_1", FieldOf("vector_field_2"))
 func EuclideanDistance(vector1 any, vector2 any) Expr {
 	return newBaseFunction("euclidean_distance", []Expr{asFieldExpr(vector1), asVectorExpr(vector2)})
 }
 
 // VectorLength creates an expression that calculates the length of a vector.
 //   - exprOrFieldPath can be a field path string, [FieldPath] or [Expr].
-//
-// Example:
-//
-//	// Calculate the length of a vector field.
-//	VectorLength("vector_field")
 func VectorLength(exprOrFieldPath any) Expr {
 	return newBaseFunction("vector_length", []Expr{asFieldExpr(exprOrFieldPath)})
 }
